Move program to separate module
This helps with the semantics of what each module accomplishes.
